% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele1.R
\name{allele1}
\alias{allele1}
\title{Allele plot for SNPs, alleles and allele pairs}
\usage{
allele1(scan_apr, coefs, coefs36, scan_pat, map, haplo)
}
\arguments{
\item{scan_apr}{Object of class \code{\link[qtl2scan]{scan1}} with allele scan.}

\item{coefs}{Object of class \code{\link[qtl2scan]{scan1coef}}.}

\item{coefs36}{Object of class \code{\link[qtl2scan]{scan1coef}} with allele pair coefficients.}

\item{scan_pat}{Object of class \code{\link{scan_pattern}}.}

\item{map}{Genome map.}

\item{haplo}{Haplotype allele letter to compare for number of copies.}
}
\value{
Table with allele effects across sources.
}
\description{
Create table of alleles for various model fits.
}
