% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoprob_to_patternprob.R
\name{genoprob_to_patternprob}
\alias{genoprob_to_patternprob}
\title{Collapse genoprob according to pattern}
\usage{
genoprob_to_patternprob(probs1, sdp)
}
\arguments{
\item{probs1}{object of class \code{\link[qtl2scan]{calc_genoprob}}}

\item{sdp}{SNP distribution pattern}
}
\value{
object of class \code{\link[qtl2scan]{calc_genoprob}}
}
\description{
Collapse genoprob according to pattern
}
\examples{
\dontrun{genoprob_to_patternprob(probs1, pattern_sets)}
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
