% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_allele1.R
\name{ggplot_allele1}
\alias{ggplot_allele1}
\alias{autoplot.allele1}
\title{Plot allele1 object}
\usage{
ggplot_allele1(
  x,
  scan1_object = NULL,
  map = NULL,
  pos = NULL,
  trim = TRUE,
  legend.position = "none",
  ...
)

\method{autoplot}{allele1}(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link{allele1}}.}

\item{scan1_object}{Optional object of class \code{\link[qtl2]{scan1}} to find peak.}

\item{map}{Genome map (required if \code{scan1_object} present).}

\item{pos}{Genome position in Mbp (supercedes \code{scan1_object}).}

\item{trim}{If \code{TRUE}, trim extreme alleles.}

\item{legend.position}{Legend position (default is \code{"none"}).}

\item{...}{Other parameters ignored.}
}
\value{
object of class \code{\link[ggplot2]{ggplot}}
}
\description{
Plot alleles for haplotype, diplotype and top patterns and genome position.
}
