% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_mediate1.R
\name{ggplot_mediate1}
\alias{ggplot_mediate1}
\alias{autoplot.mediate1}
\title{Plot mediate1 object}
\usage{
ggplot_mediate1(x, frame = FALSE, ...)

\method{autoplot}{mediate1}(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link[qtl2pattern]{allele1}}.}

\item{frame}{If \code{TRUE}, enable frames for \code{\link[plotly]{ggplotly}}}

\item{...}{Other parameters ignored.}
}
\description{
Plot mediation by expression phenotypes in region.
}
