% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_merge_feature.R
\name{ggplot_merge_feature}
\alias{ggplot_merge_feature}
\alias{autoplot.merge_feature}
\title{Plot of merge_feature object}
\usage{
ggplot_merge_feature(x, pheno, plot_by = c("pattern", "consequence"), ...)

\method{autoplot}{merge_feature}(x, ...)
}
\arguments{
\item{x}{of class \code{merge_feature}}

\item{pheno}{name of phenotype to be plotted}

\item{plot_by}{element to plot by (one of \code{c("pattern","consequence")})}

\item{...}{other arguments not used}
}
\value{
ggplot2 object
}
\description{
Plot of merge_feature object
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
