% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scan1pattern.R
\name{ggplot_scan1pattern}
\alias{ggplot_scan1pattern}
\alias{autoplot.scan1pattern}
\title{Plot scan pattern usign ggplot2}
\usage{
ggplot_scan1pattern(
  x,
  map,
  plot_type = c("lod", "coef", "coef_and_lod"),
  patterns = x$patterns$founders,
  columns = 1:3,
  min_lod = 3,
  lodcolumn = seq_along(patterns),
  facet = "pheno",
  ...
)

\method{autoplot}{scan1pattern}(x, ...)
}
\arguments{
\item{x}{object of class \code{\link{scan1pattern}}}

\item{map}{A list of vectors of marker positions, as produced by
\code{\link[qtl2]{insert_pseudomarkers}}.}

\item{plot_type}{type of plot from \code{c("lod","coef")}}

\item{patterns}{allele patterns to plot (default all)}

\item{columns}{columns used for coef plot}

\item{min_lod}{minimum LOD peak for contrast to be retained}

\item{lodcolumn}{columns used for scan1 plot (default all \code{patterns})}

\item{facet}{Plot facets if multiple phenotypes and patterns provided (default = \code{"pheno"}).}

\item{...}{additional parameters}
}
\value{
object of class \code{\link[ggplot2]{ggplot}}
}
\description{
Plot scan pattern usign ggplot2
}
