% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot.R
\name{hotspot}
\alias{hotspot}
\title{Hotspots for phenotypes}
\usage{
hotspot(map, peaks, peak_window = 1, minLOD = 5.5)
}
\arguments{
\item{map}{list of genetic maps}

\item{peaks}{data frame of peak information}

\item{peak_window}{half-width of peak window in Mbp}

\item{minLOD}{minimum LOD to include in count}
}
\value{
object of class hotspot as list of \code{\link[qtl2scan]{scan1}} and \code{map} objects.
}
\description{
Count hotspots by pheno_group and pheno_type.
}
\examples{
\dontrun{hotspot <- function(map, peaks, peak_window, minLOD)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
