% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listof_scan1coef.R
\name{listof_scan1coef}
\alias{listof_scan1coef}
\alias{summary_listof_scan1coef}
\alias{summary.listof_scan1coef}
\alias{summary_scan1coef}
\alias{summary.scan1coef}
\alias{subset_listof_scan1coef}
\alias{subset.listof_scan1coef}
\alias{[.listof_scan1coef}
\title{List of scan1coef objects}
\usage{
listof_scan1coef(
  probs,
  phe,
  K = NULL,
  covar = NULL,
  blups = FALSE,
  center = FALSE
)

summary_listof_scan1coef(
  object,
  scan1_object,
  map,
  coef_names = dimnames(object[[1]])[[2]],
  center = TRUE,
  ...
)

\method{summary}{listof_scan1coef}(object, ...)

summary_scan1coef(object, scan1_object, map, ...)

\method{summary}{scan1coef}(object, ...)

subset_listof_scan1coef(x, elements, ...)

\method{subset}{listof_scan1coef}(x, ...)

\method{[}{listof_scan1coef}(x, ...)
}
\arguments{
\item{probs}{genotype probabilities object for one chromosome from \code{\link[qtl2]{calc_genoprob}}}

\item{phe}{data frame of phenotypes}

\item{K}{list of length 1 with kinship matrix}

\item{covar}{matrix of covariates}

\item{blups}{Create BLUPs if \code{TRUE}}

\item{center}{center coefficients if \code{TRUE}}

\item{object}{object of class \code{listof_scan1coef}}

\item{scan1_object}{object from \code{scan1}}

\item{map}{A list of vectors of marker positions, as produced by
\code{\link[qtl2]{insert_pseudomarkers}}.}

\item{coef_names}{names of effect coefficients (default is all coefficient names)}

\item{...}{ignored}

\item{x}{object of class \code{listof_scan1coef}}

\item{elements}{indexes or names of list elements in x}
}
\value{
object of class \code{listof_scan1coef}
}
\description{
Create a list of scan1coef objects using \code{\link[qtl2]{scan1coef}}.

Summary of object of class \code{\link{listof_scan1coef}}, which is a list of objects of class \code{scan1coef}.

Summary of object of class \code{\link{listof_scan1coef}}, which is a list of objects of class \code{scan1coef}.

Subset of object of class \code{\link{listof_scan1coef}}, which is a list of objects of class \code{scan1coef}.
}
\examples{
dirpath <- "https://raw.githubusercontent.com/rqtl/qtl2data/master/DOex"

# Read DOex example cross from 'qtl2data'
DOex <- qtl2::read_cross2(file.path(dirpath, "DOex.zip"))
DOex <- subset(DOex, chr = "2")

# Calculate genotype and allele probabilities
pr <- qtl2::calc_genoprob(DOex, error_prob=0.002)
apr <- qtl2::genoprob_to_alleleprob(pr)

# Run scan1coef or scan1blup on all phenotypes, returning a list of \code{\link[qtl2]{scan1coef}} objects
out <- listof_scan1coef(apr, DOex$pheno, center = TRUE)

# Summary of coefficients at scan peak
scan_pr <- qtl2::scan1(pr, DOex$pheno)
summary(out, scan_pr, DOex$pmap)

plot(out[[1]], DOex$pmap, columns = LETTERS[1:8])

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
