% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate1.R
\name{mediate1}
\alias{mediate1}
\title{Mediate phenotype with nearby expression phenotypes}
\usage{
mediate1(chr_id, pos_Mbp, window_Mbp, phe_df, cov_mx = NULL, probs_obj,
  kinship = NULL, map, verbose = FALSE, region = FALSE)
}
\arguments{
\item{chr_id}{Chromosome identifier.}

\item{pos_Mbp}{Position in megabase pairs.}

\item{window_Mbp}{Window width to scan for expression traits.}

\item{phe_df}{Data frame with phenotype.}

\item{cov_mx}{Covariate matrix}

\item{probs_obj}{Object from \code{\link[qtl2feather]{feather_genoprob}}}

\item{kinship}{Kinship matrix list.}

\item{map}{Physical map.}

\item{verbose}{Show expression trait number as calculated if \code{TRUE}}

\item{region}{Mediate over whole region if \code{TRUE}}

\item{datapath}{Directory name for expression traits objects.}
}
\description{
Relies on user-provided \code{query_mrna}.
}
