% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_scan1.R
\name{modify_scan1}
\alias{modify_scan1}
\title{Modify scan1 lod matrix.}
\usage{
modify_scan1(scan1_object, newlod)
}
\arguments{
\item{scan1_object}{object of class \code{\link[qtl2scan]{scan1}}}

\item{newlod}{matrix of new LOD values}
}
\description{
Modify scan1 lod matrix.
}
