% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_diplos.R
\name{pattern_diplos}
\alias{pattern_diplos}
\alias{pattern_haplos}
\title{Extract pattern of diplotypes}
\usage{
pattern_diplos(sdp, haplos, diplos, cont = NULL)

pattern_haplos(sdp, haplos)
}
\arguments{
\item{sdp}{vector of sdp from \code{\link{top_snps_pattern}}}

\item{haplos}{vector of haplotype names}

\item{diplos}{vector of diplotype names}

\item{cont}{vector of types of contrasts (\code{NULL} or from \code{c("add","dom","b6r","b6d")})}
}
\value{
matrix of diplotype patterns

matrix of haplotype patterns
}
\description{
Extract pattern of diplotypes

Extract pattern of haplotypes
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
