% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_label.R
\name{pattern_label}
\alias{pattern_label}
\title{Turn genotype probabilities into labels}
\usage{
pattern_label(genos, sdp = NULL, allele = TRUE)
}
\arguments{
\item{sdp}{SNP distribution pattern for plot colors}

\item{allele}{Driver has alleles if \code{TRUE}, otherwise allele pairs.}

\item{genomat}{matrix of genotype probabilities at locus}
}
\description{
Turn genotype probabilities into labels
}
