% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_region.R
\name{pheno_region}
\alias{pheno_region}
\alias{expr_region}
\title{Create list with phenotypes in region}
\usage{
pheno_region(
  chr_id,
  start_val,
  end_val,
  covar,
  map,
  peaks,
  analyses,
  pheno_data,
  drivers = 2
)

expr_region(
  chr_id,
  start_val,
  end_val,
  covar,
  map,
  project_dir,
  drivers = 2,
  query_mrna = create_mrna_query_func(project_dir)
)
}
\arguments{
\item{chr_id, start_val, end_val}{chromosome and start and end value}

\item{covar}{covariate data frame}

\item{map}{list or vector of map positions}

\item{peaks}{table of peaks}

\item{analyses}{table of analyses}

\item{pheno_data}{matrix of phenotype data}

\item{drivers}{number of drivers (1 or 2; default is 2)}

\item{project_dir}{project directory with mRNA data in subdirector \code{RNAseq}}

\item{query_mrna}{query routine for mRNA data (see \code{\link{create_mrna_query_func}})}
}
\description{
Create list with phenotypes in region

Create list with expression phenotypes in region
}
\seealso{
\code{\link{create_mrna_query_func}}
}
