% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_mediator.R
\name{pull_mediator}
\alias{pull_mediator}
\title{Pull mediator and SDP from mediate object}
\usage{
pull_mediator(mediate_obj, med_ls, patterns, med_name = med_names[1])
}
\arguments{
\item{mediate_obj}{object from \code{\link[CausalMST]{mediation_test}}}

\item{med_ls}{object from \code{link{pheno_region}}}

\item{patterns}{data frame \code{summary} of \code{\link{tob_snps_all}} object}

\item{med_name}{name of mediator to pull}
}
\description{
Pull mediator and SDP from mediate object
}
