% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fast.R
\name{read_fast}
\alias{read_fast}
\title{Read fast database with possible rownames}
\usage{
read_fast(datapath, columns = NULL, rownames = TRUE, fast = c("feather",
  "fst"))
}
\arguments{
\item{datapath}{character string path to database}

\item{columns}{names or indexes for columns to be extracted}

\item{rownames}{use first column of rownames if \code{TRUE} (can supply column number)}

\item{fast}{one of \code{"feather"} or \code{"fst"}}
}
\description{
Read fast database with format \code{feather} or \code{fst}. Use first column
of database as rownames if desired. R/qtl2 routines assume data frames have
rownames to use to align individuals.
}
\seealso{
\code{\link[fst]{read_fst}}, \code{\link[feather]{read_feather}}
}
