% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1pattern.R
\name{scan1pattern}
\alias{scan1pattern}
\alias{summary.scan1pattern}
\title{Genome scan by pattern set}
\usage{
scan1pattern(
  probs1,
  phe,
  K = NULL,
  covar = NULL,
  map,
  patterns,
  condense_patterns = TRUE,
  blups = FALSE,
  do_scans = TRUE
)

\method{summary}{scan1pattern}(object, map, ...)
}
\arguments{
\item{probs1}{object of class \code{\link[qtl2]{calc_genoprob}}}

\item{phe}{data frame with one phenotype}

\item{K}{kinship matrix}

\item{covar}{covariate matrix}

\item{map}{genome map}

\item{patterns}{data frame of pattern information}

\item{condense_patterns}{remove snp_action from contrasts if TRUE}

\item{blups}{Create BLUPs if \code{TRUE}}

\item{do_scans}{Do scans if \code{TRUE}.}

\item{object}{object of class \code{\link{scan1pattern}}}

\item{...}{additional parameters passed on to other methods}
}
\value{
List containing:
\itemize{
\item{patterns} Data frame of summary for top patterns (column \code{founders} has pattern)
\item{dip_set} Diplotype sets for contrasts
\item{group} Group for each founder pattern
\item{scan} Object of class \code{\link[qtl2]{scan1}}.
\item{coef} Object of class \code{listof_scan1coef}. See package 'qtl2ggplot'.
}
}
\description{
Genome scan by pattern set
}
\examples{
#\donttest{
# Long elapsed time due to calc_genoprob.

dirpath <- "https://raw.githubusercontent.com/rqtl/qtl2data/master/DOex"

# Read DOex example cross from 'qtl2data'
DOex <- subset(qtl2::read_cross2(file.path(dirpath, "DOex.zip")), chr = "2")

# Download genotype probabilities
tmpfile <- tempfile()
download.file(file.path(dirpath, "DOex_genoprobs.rds"), tmpfile, quiet=TRUE)
pr <- subset(readRDS(tmpfile), chr = "2")
unlink(tmpfile)

# Download SNP info for DOex from web and read as RDS.
tmpfile <- tempfile()
download.file(file.path(dirpath, "c2_snpinfo.rds"), tmpfile, quiet=TRUE)
snpinfo <- readRDS(tmpfile)
unlink(tmpfile)
snpinfo <- dplyr::rename(snpinfo, pos = pos_Mbp)

# Convert to SNP probabilities
snpinfo <- qtl2::index_snps(DOex$pmap, snpinfo)
snppr <- qtl2::genoprob_to_snpprob(pr, snpinfo)

# Scan SNPs
scan_snppr <- qtl2::scan1(snppr, DOex$pheno)
top_snps_tbl <- top_snps_pattern(scan_snppr, snpinfo)

# Summarize to find top patterns
patterns <- dplyr::arrange(summary(top_snps_tbl), dplyr::desc(max_lod))

# Scan using patterns.
scan_pat <- scan1pattern(pr, DOex$pheno, map = DOex$pmap, patterns = patterns)

# Summary of scan1pattern.
summary(scan_pat, DOex$pmap)
#}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
