% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_pattern.R
\name{scan_pattern}
\alias{scan_pattern}
\alias{summary.scan_pattern}
\title{Genome scan by pattern set}
\usage{
scan_pattern(probs1, phe, K = NULL, covar = NULL, map, patterns,
  haplos = NULL, diplos = NULL, condense_patterns = TRUE, blups = FALSE,
  do_scans = TRUE)

\method{summary}{scan_pattern}(object, map, ...)
}
\arguments{
\item{probs1}{object of class \code{\link[qtl2geno]{calc_genoprob}}}

\item{phe}{data frame with one phenotype}

\item{K}{kinship matrix}

\item{covar}{covariate matrix}

\item{map}{genome map}

\item{patterns}{data frame of pattern information}

\item{haplos}{vector of haplotype names}

\item{diplos}{vector of diplotype names}

\item{condense_patterns}{remove snp_action from contrasts if TRUE}

\item{blups}{Create BLUPs if \code{TRUE}}

\item{do_scans}{Do scans if \code{TRUE}.}

\item{object}{object of class \code{\link{scan_pattern}}}
}
\value{
List containing:
\itemize{
\item{patterns} Data frame of summary for top patterns (column \code{founders} has pattern)
\item{dip_set} Diplotype sets for contrasts
\item{group} Group for each founder pattern
\item{scan} Object of class \code{\link[qtl2scan]{scan1}}.
\item{coef} Object of class \code{\link{listof_scan1coef}}.
}
}
\description{
Genome scan by pattern set
}
\examples{
\dontrun{scan_pattern(probs, phe, K, covar, patterns, haplos, diplos)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
