% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_to_pattern.R
\name{sdp_to_pattern}
\alias{sdp_to_pattern}
\alias{sdp_to_logical}
\title{Convert sdp to pattern}
\usage{
sdp_to_pattern(sdp, haplos)

sdp_to_logical(sdp, haplos)
}
\arguments{
\item{sdp}{vector of sdp values}

\item{haplos}{letter codes for haplotypes (required)}
}
\value{
vector of letter patterns
}
\description{
Convert strain distribution pattern (sdp) to letter pattern.
}
\examples{
\dontrun{sdp_to_pattern(snp, haplos)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
