% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpscan_pattern.R
\name{snpscan_pattern}
\alias{snpscan_pattern}
\alias{summary.snpscan_pattern}
\alias{plot.snpscan_pattern}
\alias{plot_snpscan}
\title{Plot coefficients by pattern set}
\usage{
snpscan_pattern(probs1, phe, K, covar, pattern)

\method{summary}{snpscan_pattern}(object, ...)

\method{plot}{snpscan_pattern}(x, title = "", colors = NULL,
  col_names = NULL, color_lod = 1, coef = TRUE, lod = TRUE,
  center = TRUE, max_coef = 5, ...)

plot_snpscan(probs1, phe, K, covar, patterns, haplos, diplos)
}
\arguments{
\item{probs1}{object of class \code{\link[qtl2scan]{calc_genoprob}}}

\item{phe}{data frame with one phenotype}

\item{K}{kinship matrix}

\item{covar}{covariate matrix}

\item{object}{object of class \code{\link{snpscan_pattern}}}

\item{...}{additional parameters}

\item{x}{object of class \code{\link{snpscan_pattern}}}

\item{title}{title of plot (default "")}

\item{colors}{colors for plot (default \code{link[qtl2ggplot]{CCcolors}})}

\item{col_names}{names of colors (default from patterns)}

\item{color_lod}{color of LOD plot line}

\item{center}{center coefficients if \code{TRUE} (default)}

\item{max_coef}{maximum absolute coefficient for plotting (default 5)}

\item{patterns}{data frame of pattern information}

\item{haplos}{vector of haplotype names}

\item{diplos}{vector of diplotype names}

\item{patternl}{list of patterns}

\item{probs2}{genotype probabilities for pattern set}

\item{title}{title for plots (default "")}

\item{col_names}{names of colors (default from patterns)}

\item{probs1}{object of class \code{\link[qtl2scan]{calc_genoprob}}}

\item{phe}{data frame with one phenotype}

\item{K}{kinship matrix}

\item{covar}{covariate matrix}

\item{patterns}{data frame of pattern information}
}
\value{
ggplot2 object

object of class \code{snpscan_pattern}

invisible probs collapsed to pattern

ggplot2 object

invisible probs collapsed to pattern

ggplot2 object
}
\description{
Plot coefficients by pattern set

summary of snpscan_pattern object

Plot of snpscan_pattern object

Plot coefficients by pattern set
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
