% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topsnp_pattern.R
\name{topsnp_pattern}
\alias{topsnp_pattern}
\alias{summary.topsnp_pattern}
\alias{plot.topsnp_pattern}
\title{Fine Mapping scans by allele pattern}
\usage{
topsnp_pattern(out_lmm_snps, proband, drop = 1.5, min_lod = 3)

\method{summary}{topsnp_pattern}(object, type = c("best", "common"), ...)

\method{plot}{topsnp_pattern}(x, ..., title = "", group = c("pheno",
  "pattern"))
}
\arguments{
\item{out_lmm_snps}{output of linear mixed model for \code{phename} (see \code{\link[qtl2scan]{scan1}})}

\item{proband}{character string being one or more of the phenotypes in \code{out_lmm}}

\item{drop}{include all SNPs within \code{drop} of max LOD (default 1.5)}

\item{min_lod}{minimum LOD retained (default 1.5)}

\item{object}{of class \code{topsnp_pattern}}

\item{type}{type of summary}

\item{...}{other arguments not used}

\item{x}{of class \code{topsnp_pattern}}

\item{title}{title of plot (default "")}

\item{group}{group plot by one of c("pheno","pattern")}

\item{...}{other arguments not used}
}
\value{
table of top_snps at maximum lod for \code{pattern}

table summary

ggplot2 object
}
\description{
Separate fine mapping scans by allele pattern.

Summary of topsnp_pattern object

Plot of topsnp_pattern object
}
\examples{
dontrun(topsnp_pattern(out_lmm_snps, proband))

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{hplot}
\keyword{utilities}
