% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele1.R
\name{allele1}
\alias{allele1}
\title{Allele plot for SNPs, alleles and allele pairs}
\usage{
allele1(phe_df = NULL, cov_mx = NULL, probD = NULL, map = NULL,
  K_chr = NULL, patterns = NULL, alt = NULL, trim = TRUE, ...)
}
\arguments{
\item{phe_df}{data frame with one phenotype}

\item{cov_mx}{covariate matrix}

\item{probD}{object of class \code{\link[qtl2geno]{calc_genoprob}}}

\item{map}{list of genome maps}

\item{K_chr}{kinship matrix}

\item{patterns}{data frame of pattern information}

\item{alt}{Haplotype allele letter(s) for alternative to reference.}

\item{trim}{If \code{TRUE}, trim extreme alleles.}

\item{...}{additional parameters}

\item{scanH}{Object of class \code{\link[qtl2scan]{scan1}} with allele scan.}

\item{coefH}{Object of class \code{\link[qtl2scan]{scan1coef}}.}

\item{coefD}{Object of class \code{\link[qtl2scan]{scan1coef}} with allele pair coefficients.}

\item{scan_pat}{Object of class \code{\link{scan_pattern}}.}
}
\value{
Table with allele effects across sources.
}
\description{
Create table of alleles for various model fits.
}
