% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covar.R
\name{get_covar}
\alias{get_covar}
\title{Get covariates}
\usage{
get_covar(covar, analyses_tbl)
}
\arguments{
\item{covar}{matrix of all covariates}

\item{analyses_tbl}{table of analyses setups}
}
\value{
matrix of covariates
}
\description{
Get covariates using \code{analyses_tbl} filtered by \code{phename_output}.
}
\examples{
\dontrun{get_covar(covar, datapath)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
