% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_label.R
\name{pattern_label}
\alias{pattern_label}
\alias{pattern_sdp}
\title{Turn genotype probabilities into labels}
\usage{
pattern_label(genos, allele = TRUE)

pattern_sdp(label, sdp = NULL, geno_names = sort(unique(label)))
}
\arguments{
\item{genos}{matrix of genotype probabilities at locus}

\item{allele}{Driver has alleles if \code{TRUE}, otherwise allele pairs.}

\item{label}{character string from \code{\link{pattern_label}}}

\item{sdp}{SNP distribution pattern for plot colors}

\item{geno_names}{unique genotype names (alleles or allele pairs)}
}
\value{
character vector of genotype names.
}
\description{
Turn genotype probabilities into labels
}
