% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_pattern.R
\name{scan_pattern}
\alias{scan_pattern}
\alias{summary.scan_pattern}
\title{Genome scan by pattern set}
\usage{
scan_pattern(probs1, phe, K, covar, patterns, haplos = NULL, diplos = NULL,
  condense_patterns = TRUE)

\method{summary}{scan_pattern}(object, ...)
}
\arguments{
\item{probs1}{object of class \code{\link[qtl2scan]{calc_genoprob}}}

\item{phe}{data frame with one phenotype}

\item{K}{kinship matrix}

\item{covar}{covariate matrix}

\item{patterns}{data frame of pattern information}

\item{haplos}{vector of haplotype names}

\item{diplos}{vector of diplotype names}

\item{condense_patterns}{remove snp_action from contrasts if TRUE}

\item{object}{object of class \code{\link{scan_pattern}}}
}
\value{
ggplot2 object
}
\description{
Genome scan by pattern set
}
\examples{
\dontrun{scan_pattern(probs, phe, K, covar, patterns, haplos, diplos)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
