% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpprob_collapse.R
\name{snpprob_collapse}
\alias{snpprob_collapse}
\title{Collapse genoprob according to pattern}
\usage{
snpprob_collapse(
  snpprobs,
  action = c("additive", "add+dom", "non-add", "recessive", "dominant")
)
}
\arguments{
\item{snpprobs}{object of class \code{\link[qtl2]{calc_genoprob}}}

\item{action}{SNP gene action type}
}
\value{
object of class \code{\link[qtl2]{calc_genoprob}}
}
\description{
Collapse genoprob according to pattern
}
\examples{
\dontrun{snpprob_collapse(snpprobs, snpsets)}
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
